feat: less config and add umi-plugin-react

+++ scripts/build.js
@@ -3,7 +3,13 @@ const babel = require('@babel/core');
 const through = require('through2');
 const chalk = require('chalk');
 const rimraf = require('rimraf');
-const { readdirSync, readFileSync, writeFileSync, existsSync } = require('fs');
+const {
+  readdirSync,
+  readFileSync,
+  writeFileSync,
+  existsSync,
+  statSync,
+} = require('fs');
 const { join } = require('path');
 const chokidar = require('chokidar');
 const shell = require('shelljs');
@@ -121,20 +127,22 @@ function build() {
       watcher.on('all', (event, fullPath) => {
         if (!existsSync(fullPath)) return;
         const relPath = fullPath.replace(`${cwd}/packages/${pkg}/src/`, '');
-        const content = readFileSync(fullPath, 'utf-8');
-        try {
-          const code = transform({
-            content,
-            path: fullPath,
-          });
-          writeFileSync(
-            join(cwd, 'packages', pkg, 'lib', relPath),
-            code,
-            'utf-8',
-          );
-        } catch (e) {
-          console.log(chalk.red('Compiled failed.'));
-          console.log(chalk.red(e.message));
+        if (statSync(fullPath).isFile()) {
+          const content = readFileSync(fullPath, 'utf-8');
+          try {
+            const code = transform({
+              content,
+              path: fullPath,
+            });
+            writeFileSync(
+              join(cwd, 'packages', pkg, 'lib', relPath),
+              code,
+              'utf-8',
+            );
+          } catch (e) {
+            console.log(chalk.red('Compiled failed.'));
+            console.log(chalk.red(e.message));
+          }
         }
       });
     }
